Create SBOM using Syft (#134)
All container images and binaries now ship with a software bill of material in spdx-json format. The SBOMs for the binaries are attached to the GitHub Releases. The SBOMs for each container image tag are attached as attestations. 
Each SBOM is signed and can be verified using the cosign.pub key.

This commit also changes the provenance extension from .json to .att (Attestation)

Closes #61 #128